+++ modules/postgresql/versions.tf
@@ -36,10 +36,10 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-sql-db:postgresql/v20.0.0"
+    module_name = "blueprints/terraform/terraform-google-sql-db:postgresql/v20.2.0"
   }
   provider_meta "google-beta" {
-    module_name = "blueprints/terraform/terraform-google-sql-db:postgresql/v20.0.0"
+    module_name = "blueprints/terraform/terraform-google-sql-db:postgresql/v20.2.0"
   }
 
 }