+++ modules/safer_mysql/versions.tf
@@ -25,7 +25,7 @@ terraform {
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-sql-db:safer_mysql/v20.0.0"
+    module_name = "blueprints/terraform/terraform-google-sql-db:safer_mysql/v20.2.0"
   }
 
 }